+++ apps/fluent-tester/src/E2E/CheckboxExperimental/specs/ExperimentalCheckbox.spec.win.ts
@@ -10,7 +10,7 @@ import { PAGE_TIMEOUT, BOOT_APP_TIMEOUT, CHECKBOX_A11Y_ROLE } from '../../common
 describe('Experimental Checkbox Testing Initialization', () => {
   it('Wait for app load', () => {
     NavigateAppPage.waitForPageDisplayed(BOOT_APP_TIMEOUT);
-    expect(NavigateAppPage.isPageLoaded()).toBeTruthy();
+    expect(NavigateAppPage.isPageLoaded()).toBeTruthy(NavigateAppPage.ERRORMESSAGE_APPLOAD);
   });
 
   it('Click and navigate to Experimental Checkbox test page', () => {
@@ -22,7 +22,8 @@ describe('Experimental Checkbox Testing Initialization', () => {
     NavigateAppPage.clickAndGoToCheckboxExperimentalPage();
     ExperimentalCheckboxPageObject.waitForPageDisplayed(PAGE_TIMEOUT);
 
-    expect(ExperimentalCheckboxPageObject.isPageLoaded()).toBeTruthy();
+    expect(ExperimentalCheckboxPageObject.isPageLoaded()).toBeTruthy(ExperimentalCheckboxPageObject.ERRORMESSAGE_PAGELOAD);
+    expect(ExperimentalCheckboxPageObject.didAssertPopup()).toBeFalsy(ExperimentalCheckboxPageObject.ERRORMESSAGE_ASSERT); // Ensure no asserts popped up
   });
 });
 
@@ -30,22 +31,28 @@ describe('Experimental Checkbox Accessibility Testing', () => {
   it('Experimental Checkbox - Validate accessibilityRole is correct', () => {
     ExperimentalCheckboxPageObject.scrollToTestElement();
     ExperimentalCheckboxPageObject.waitForPrimaryElementDisplayed(PAGE_TIMEOUT);
+
     expect(ExperimentalCheckboxPageObject.getAccessibilityRole()).toEqual(CHECKBOX_A11Y_ROLE);
+    expect(ExperimentalCheckboxPageObject.didAssertPopup()).toBeFalsy(ExperimentalCheckboxPageObject.ERRORMESSAGE_ASSERT);
   });
 
   it('Experimental Checkbox - Set accessibilityLabel', () => {
     ExperimentalCheckboxPageObject.scrollToTestElement();
     ExperimentalCheckboxPageObject.waitForPrimaryElementDisplayed(PAGE_TIMEOUT);
+
     expect(ExperimentalCheckboxPageObject.getAccessibilityLabel(ComponentSelector.Primary)).toEqual(
       EXPERIMENTAL_CHECKBOX_ACCESSIBILITY_LABEL,
     );
+    expect(ExperimentalCheckboxPageObject.didAssertPopup()).toBeFalsy(ExperimentalCheckboxPageObject.ERRORMESSAGE_ASSERT);
   });
 
   it('Experimental Checkbox - Do not set accessibilityLabel -> Default to Checkbox label', () => {
     ExperimentalCheckboxPageObject.scrollToTestElement();
     ExperimentalCheckboxPageObject.waitForPrimaryElementDisplayed(PAGE_TIMEOUT);
+
     expect(ExperimentalCheckboxPageObject.getAccessibilityLabel(ComponentSelector.Secondary)).toEqual(
       EXPERIMENTAL_CHECKBOX_TEST_COMPONENT_LABEL,
     );
+    expect(ExperimentalCheckboxPageObject.didAssertPopup()).toBeFalsy(ExperimentalCheckboxPageObject.ERRORMESSAGE_ASSERT);
   });
 });