+++ packages/components/ContextualMenu/src/ContextualMenu.types.ts
@@ -1,6 +1,8 @@
 import { ICalloutProps, ICalloutTokens } from '@fluentui-react-native/callout';
 import { IViewProps } from '@fluentui-react-native/adapters';
 import { IRenderData } from '@uifabricshared/foundation-composable';
+import { ScrollViewProps } from 'react-native';
+import { FocusZoneProps } from '@fluentui-react-native/focus-zone';
 
 export const contextualMenuName = 'ContextualMenu';
 export interface ContextualMenuContext {
@@ -63,6 +65,8 @@ export interface ContextualMenuProps extends IViewProps, Omit<ICalloutProps, 'se
 export type ContextualMenuSlotProps = {
   root: ContextualMenuProps;
   container: IViewProps;
+  scrollView: ScrollViewProps;
+  focusZone?: FocusZoneProps; // macOS only
 };
 
 export type ContextualMenuRenderData = IRenderData<ContextualMenuSlotProps, ContextualMenuState>;